Reports
Added title to repot so name can be for organizational and tile can
print.  Started report copy.  Report filter query working, plumb value
to UI, and post back selection so it's a parameter to an SQL query.
Added ACE Editor for HTML and SQL editing.  CSS for reports and ACE
Editor.